refactor: SwiftLint 룰에 맞게 수정.
