+++ CloudNotes/CloudNotes/Controllers/CloudMemoSplitViewController.swift
@@ -33,7 +33,9 @@ class CloudMemoSplitViewController: UISplitViewController {
 }
 
 extension CloudMemoSplitViewController: UISplitViewControllerDelegate {
-    func splitViewController(_ splitViewController: UISplitViewController, collapseSecondary secondaryViewController: UIViewController, onto primaryViewController: UIViewController) -> Bool {
+    func splitViewController(_ splitViewController: UISplitViewController,
+                             collapseSecondary secondaryViewController: UIViewController,
+                             onto primaryViewController: UIViewController) -> Bool {
         return true
     }
 }
